// src/types/partials/QuestionnaireResponseItem.ts
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TUri } from '../simpleTypes/Uri';
import { TQuestionnaireResponseAnswer } from '../partials/QuestionnaireResponseAnswer';

export type TQuestionnaireResponseItem = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    linkId: String;
    definition?: TUri;
    text?: String;
    answer?: TQuestionnaireResponseAnswer[];
    item?: TQuestionnaireResponseItem[];
};